pyproject.toml
[tool.poetry]
name = "online-cinema-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.37"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
black = "^25.1.0"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
pillow = "^11.1.0"
boto3 = "^1.36.11"
alembic = "^1.14.1"

[tool.black]
line-length = 88
exclude = "app/database/migrations"

[tool.flake8]
max-line-length = 88
ignore = "F821, W503"
max-complexity = 18
exclude = ".venv,venv,migrations"
count = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
